Update to use 64-bit
This commit modifies CMake files and data types to build in 64-bit
modes. These changes will break 32-bit build. Note that qsort is
Intel specific, so it may need to be addressed for GNU.

+++ dsm2/src/hdf_tidefile/dsm2_tidefile_input_storage_hydro_comp_point.cpp
@@ -253,7 +253,7 @@ void hydro_comp_point_number_rows_hdf5_f(const hid_t *file_id, hsize_t* nrecords
 
     
 /** get one row worth of information from the buffer */
-void hydro_comp_point_query_from_buffer_f(size_t* row, 
+void hydro_comp_point_query_from_buffer_f(int32_t* row, 
                         int * a_comp_index,int * a_channel,double * a_distance, int * ierror
                         )
 {
