+++ dsm2/src/hdf_tidefile/dsm2_tidefile_input_storage_stage_boundaries.cpp
@@ -261,7 +261,7 @@ void stage_boundaries_number_rows_hdf5_f(const hid_t *file_id, hsize_t* nrecords
 
     
 /** get one row worth of information from the buffer */
-void stage_boundaries_query_from_buffer_f(size_t* row, 
+void stage_boundaries_query_from_buffer_f(int32_t* row, 
                          char a_name[32],int * a_int_node_no,int * a_ext_node_no, int * ierror, 
               int name_len
                         )
